👱🏓 Checkpoint ./komutlar/dm-spammer.js:26490528/233 ./komutlar/spammer.js:26490528/111

+++ komutlar/dm-spammer.js
@@ -1,4 +1,5 @@
 const Discord = require('discord.js-selfbot');
+const db = require('quick.db')
 const ayarlar = require('../ayarlar.json')
 ///spammer by planta
  
@@ -10,7 +11,7 @@ exports.run = (client, message, args) => {
       let mesaj = args.slice(2).join(' ');
 
 if (!kullanıcı) return message.channel.send(`${basarisiz} ${message.author}, Bir kullanıcı etiketle/ID gir.`).then(x => x.delete({timeout: 5000}))
-if (isNaN(sayi)) return message.channel.send(`${basarisiz} ${message.author}, .dm-spam <kullanıcı> <sayı> <mesaj> şeklinde yazınız.`).then(x => x.delete({timeout: 5000}))
+if (isNaN(sayi)) return message.channel.send(`${basarisiz} ${message.author}, ${db.fetch(`prefix`) || ayarlar.prefix}dm-spam <kullanıcı> <sayı> <mesaj> şeklinde yazınız.`).then(x => x.delete({timeout: 5000}))
 if (mesaj.length < 1) return message.channel.send(`${basarisiz} ${message.author}, Kralım Spamlamam İçin Bişe Yazmalısınız.`);
    message.delete();
 for (var i = 0; i < sayi; i++)
